feat: studygroup의 TimeSchedule 자동 생성 시 데이터 매핑
[skip ci]
