+++ server/edusync/src/main/java/com/codestates/edusync/model/study/studygroup/utils/ScheduleConverter.java
@@ -1,5 +1,11 @@
 package com.codestates.edusync.model.study.studygroup.utils;
 
+import com.codestates.edusync.model.common.entity.TimeRange;
+import com.codestates.edusync.model.study.plancalendar.entity.TimeSchedule;
+import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
+
+import java.time.LocalDateTime;
+import java.time.Period;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -9,8 +15,8 @@ import java.util.regex.Pattern;
 public class ScheduleConverter {
 
     private static int getDayIndex(String day) {
-        List<String> daysOfWeek = Arrays.asList("일", "월", "화", "수", "목", "금", "토");
-        return daysOfWeek.indexOf(day);
+        List<String> daysOfWeek = Arrays.asList("월", "화", "수", "목", "금", "토", "일");
+        return daysOfWeek.indexOf(day)+1;
     }
 
     public static List<Integer> convertToIndex(String daysOfWeek) {
@@ -26,4 +32,58 @@ public class ScheduleConverter {
 
         return result;
     }
+
+    public static List<TimeSchedule> repeatedScheduleToScheduleListConverter(Studygroup studygroup) {
+        List<TimeSchedule> timeSchedules = new ArrayList<>();
+
+        LocalDateTime periodStart = studygroup.getDate().getStudyPeriodStart();
+        LocalDateTime periodEnd = studygroup.getDate().getStudyPeriodEnd();
+
+        LocalDateTime timeStart = studygroup.getTime().getStudyTimeStart();
+        LocalDateTime timeEnd = studygroup.getTime().getStudyTimeEnd();
+        int continueToNextDayOffset = 0;
+        if( timeStart.getHour() > timeEnd.getHour() ) {
+            continueToNextDayOffset = 1;
+        }
+
+        Period totalPeriod = Period.between(periodStart.toLocalDate(), periodEnd.toLocalDate());
+
+        List<Integer> indexOfWeeks = ScheduleConverter.convertToIndex(studygroup.getDaysOfWeek());
+        for( int offset = 0; offset <= totalPeriod.getDays(); offset++ ) {
+            LocalDateTime offsetDate = periodStart.plusDays(offset);
+
+            int currentIndexOfWeek = offsetDate.getDayOfWeek().getValue();
+            if( indexOfWeeks.contains(currentIndexOfWeek) ) {
+                TimeSchedule ts = new TimeSchedule();
+                TimeRange tr = TimeRange.builder()
+                        .studyTimeStart(
+                                LocalDateTime.of(
+                                        offsetDate.getYear(),
+                                        offsetDate.getMonth(),
+                                        offsetDate.getDayOfMonth(),
+                                        timeStart.getHour(),
+                                        timeStart.getMinute(),
+                                        timeStart.getSecond()
+                                ))
+                        .studyTimeEnd(
+                                LocalDateTime.of(
+                                        offsetDate.plusDays(continueToNextDayOffset).getYear(),
+                                        offsetDate.plusDays(continueToNextDayOffset).getMonth(),
+                                        offsetDate.plusDays(continueToNextDayOffset).getDayOfMonth(),
+                                        timeEnd.getHour(),
+                                        timeEnd.getMinute(),
+                                        timeEnd.getSecond()
+                                ))
+                        .build();
+                ts.setTime(tr);
+                ts.setTitle(studygroup.getStudyName());
+                ts.setPlatform(studygroup.getPlatform());
+                ts.setDescription(studygroup.getIntroduction());
+                timeSchedules.add(ts);
+            }
+        }
+
+        return timeSchedules;
+    }
+
 }
